backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
joblib
numpy
scikit-learn
xgboost
python-dotenv
shap
reportlab
